Don't export docker image, if all layers already exist in the remote.
This is an optimization of the `dogestry push` command that's especially
handy when pushing the same image with multiple tags.

Previously, it would export the entire image (which can be quite slow),
before figuring out that it didn't need to push any layers to the remote.
Now, it checks the layers first, and only exports if necessary.

Fixes #22.