Add command to list images on remote.
The 'list' command only shows 'REPOSITORY' and 'TAG'. While it
would be nice to show 'IMAGE ID' as well, to do so for an S3 remote
would require performing a GET for every single image.

For now, I've opted for 'list' command that returns quickly, but has
less information, rather than a slower command that returns more info.