Always update 'repositories/...' keys when pushing to S3.
It doesn't matter if the 'repositories/<image_name>/<tag_name>'
key already exists in S3. Because the contents can change, we
must always make sure to upload the new key every time we push
an image, so that the metadata for the image/tag is always up
to date.

Fixes #32.