Remove localKeysToPush() for S3 remote.
Now that we are only writing to disk files that we know we need to
upload to S3, we don't need to check again after the image export.

Before, when we were writing all files from the docker export, it
made sense to check if a file already existed before uploading.