Get rid of a bogus log message in server
Server was displaying a bogus message "Performing regular dogestry pull (one or more hosts is not running dogestry server)!" every time a new pull came in. This is because the server calls on CmdPull() which didn't have any logic to determine if we are running in server mode.

gofmt for the other file.